+++ packages/terra-framework-docs/src/terra-dev-site/test/worklist-data-grid/DefaultWorklistDataGrid.test.jsx
@@ -10,7 +10,7 @@ const DefaultWorklistDataGrid = () => {
     <React.Fragment>
       <WorklistDataGrid
         id="default-terra-worklist-data-grid"
-        columns={cols}
+        overflowColumns={cols}
         rows={rows}
         rowHeaderIndex={rowHeaderIndex}
         columnHeaderHeight="50px"