+++ packages/terra-worklist-data-grid/tests/jest/WorklistDataGrid.test.jsx
@@ -1,10 +1,11 @@
-// eslint-disable no-console
 import React from 'react';
 /* eslint-disable-next-line import/no-extraneous-dependencies */
-import { shallowWithIntl } from 'terra-enzyme-intl';
+import { mountWithIntl, shallowWithIntl } from 'terra-enzyme-intl';
 import WorklistDataGrid from '../../src/WorklistDataGrid';
 import ColumnHeader from '../../src/subcomponents/ColumnHeader';
 import Row from '../../src/subcomponents/Row';
+
+import ERRORS from '../../src/utils/constants';
 
 // Source data for tests
 const dataFile = {
@@ -40,12 +41,29 @@ const dataFile = {
     },
   ],
 };
-describe('WorklistDataGrid', () => {
-  it('verifies that the grid created is consistent with the rows and columns props', () => {
+
+beforeAll(() => {
+  jest.spyOn(console, 'error').mockImplementation();
+  jest.spyOn(console, 'warn').mockImplementation();
+});
+
+afterEach(() => {
+  console.error.mockClear(); // eslint-disable-line no-console
+  console.warn.mockClear(); // eslint-disable-line no-console
+});
+
+afterAll(() => {
+  console.error.mockRestore(); // eslint-disable-line no-console
+  console.warn.mockRestore(); // eslint-disable-line no-console
+});
+
+describe('basic grid', () => {
+  it('verifies that the grid created is consistent with the rows and overflowColumns props', () => {
     const wrapper = shallowWithIntl(
       <WorklistDataGrid
         id="test-terra-worklist-data-grid"
-        columns={dataFile.cols}
+        pinnedColumns={dataFile.cols.slice(0, 2)}
+        overflowColumns={dataFile.cols.slice(2)}
         rows={dataFile.rows}
       />,
     ).dive();
@@ -56,13 +74,12 @@ describe('WorklistDataGrid', () => {
 
     // The number of rows should match the given data.
     expect(wrapper.find(Row)).toHaveLength(dataFile.rows.length);
-
     expect(wrapper).toMatchSnapshot();
   });
 
   it('verifies the rows are created with the right props', () => {
-    const verifyRow = (rowIndex, rowComponent, data, columns) => {
-      expect(rowComponent.props.displayedColumns).toEqual(columns);
+    const verifyRow = (rowIndex, rowComponent, data, overflowColumns) => {
+      expect(rowComponent.props.displayedColumns).toEqual(overflowColumns);
       expect(rowComponent.props.hasRowSelection).toBe(false);
       expect(rowComponent.key).toEqual(data.id);
       expect(rowComponent.props.onCellSelect).toBeDefined();
@@ -75,7 +92,7 @@ describe('WorklistDataGrid', () => {
     const wrapper = shallowWithIntl(
       <WorklistDataGrid
         id="test-terra-worklist-data-grid"
-        columns={dataFile.cols}
+        overflowColumns={dataFile.cols}
         rows={dataFile.rows}
       />,
     ).dive();
@@ -87,5 +104,94 @@ describe('WorklistDataGrid', () => {
     verifyRow(2, rows.get(2), dataFile.rows[2], dataFile.cols);
 
     expect(wrapper).toMatchSnapshot();
+  });
+});
+
+describe('with pinned columns', () => {
+  it('sets pinnedColumns as pinned', () => {
+    const wrapper = mountWithIntl(
+      <WorklistDataGrid
+        id="test-terra-worklist-data-grid"
+        pinnedColumns={dataFile.cols.slice(0, 2)}
+        overflowColumns={dataFile.cols.slice(2)}
+        rows={dataFile.rows}
+      />,
+    );
+
+    const pinnedColumnHeaderCells = wrapper.find('.pinned');
+
+    expect(pinnedColumnHeaderCells).toHaveLength(2 * 4);
+  });
+
+  it('sets row selection column as pinned', () => {
+    const wrapper = mountWithIntl(
+      <WorklistDataGrid
+        id="test-terra-worklist-data-grid"
+        pinnedColumns={dataFile.cols.slice(0, 2)}
+        overflowColumns={dataFile.cols.slice(2)}
+        rows={dataFile.rows}
+        hasSelectableRows
+      />,
+    );
+
+    const pinnedColumnHeaderCells = wrapper.find('.pinned');
+
+    expect(pinnedColumnHeaderCells).toHaveLength(3 * 4);
+  });
+
+  it('pins row selection column if pinnedColumns is undefined', () => {
+    const wrapper = mountWithIntl(
+      <WorklistDataGrid
+        id="sdfdss"
+        overflowColumns={dataFile.cols}
+        rows={dataFile.rows}
+        hasSelectableRows
+      />,
+    );
+
+    const pinnedColumnHeaderCells = wrapper.find('.pinned');
+
+    expect(pinnedColumnHeaderCells).toHaveLength(1 * 4);
+    expect(console.warn).toHaveBeenCalledWith(expect.stringContaining(ERRORS.PINNED_COLUMNS_UNDEFINED)); // eslint-disable-line no-console
+  });
+});
+
+describe('Error handling - prop types', () => {
+  it('throws an error if rowHeaderIndex is not an integer', () => {
+    shallowWithIntl(
+      <WorklistDataGrid
+        id="test-terra-worklist-data-grid"
+        rows={dataFile.rows}
+        rowHeaderIndex="2"
+      />,
+    ).dive();
+
+    expect(console.error).toHaveBeenCalledWith(expect.stringContaining(ERRORS.ROW_HEADER_INDEX_NOT_AN_INTEGER)); // eslint-disable-line no-console
+  });
+
+  it('throws an error if rowHeaderIndex is not a positive integer', () => {
+    shallowWithIntl(
+      <WorklistDataGrid
+        id="test-terra-worklist-data-grid"
+        rows={dataFile.rows}
+        rowHeaderIndex={-1}
+      />,
+    ).dive();
+
+    expect(console.error).toHaveBeenCalledWith(expect.stringContaining(ERRORS.ROW_HEADER_INDEX_LESS_THAN_ZERO)); // eslint-disable-line no-console
+  });
+
+  it('throws an error if rowHeaderIndex is greater than the length of pinned columns', () => {
+    shallowWithIntl(
+      <WorklistDataGrid
+        id="test-terra-worklist-data-grid"
+        pinnedColumns={dataFile.cols.slice(0, 2)}
+        overflowColumns={dataFile.cols.slice(2)}
+        rowHeaderIndex={2}
+        rows={dataFile.rows}
+      />,
+    ).dive();
+
+    expect(console.error).toHaveBeenCalledWith(expect.stringContaining(ERRORS.ROW_HEADER_INDEX_EXCEEDS_PINNED)); // eslint-disable-line no-console
   });
 });